change max checking in doTick in game.js
moved some of the max checking to before the core part of doTick happens. This prevents the inputs being consumed even if the newCount would get set back to max at the end of doTick.
I'm not familiar with running github pages or your coding style so you might want to make some tweaks.